Value-bootstrapped search: rollout truncation + leaf-value training (negative result)
SearchPlayer gains rollout_tricks: truncate rollouts N tricks past a sim
start and score the leaf with the value head via HeartsEnv::ObserveFor(seat)
(observation from any seat perspective). SelfPlayGen --value-out emits leaf
value records (all four seats at every trick boundary, true final outcomes;
48/deal) and distill.py --leaf-data trains the value head on them through
interleaved value-only batches. T=13 verified bit-identical to full rollouts.

Result, measured (300 paired deals, K=64, cand E net): full rollout -1.79 vs
raw; T=12 +0.49 of full (noise); T=3/2/1 collapse to +4.0..+4.4 vs raw.
Retraining V on the exact leaf distribution (leaf EV 0.40 -> 0.90, policy
unharmed) does NOT recover it: T=1..3 still +4.2..+5.3.

Diagnosis: the observation is seat-visible information only, so V cannot
condition on the determinized hidden hands - it collapses the K-determinization
ensemble to one visible-info estimate, discarding exactly what the search
averages over. Truncated search degenerates to myopic lookahead with a biased
evaluator. A leaf evaluator for determinized search must SEE the sampled
hands (oracle/perfect-information value, double-dummy style) - candidate next
step. Full-rollout search remains the deployed configuration (rollout_tricks
default -1; regression-guarded).

Also: InferenceServer row counter widened to long long (32-bit Windows long
overflowed on billion-row runs).

+++ InferenceServer.hpp
@@ -100,9 +100,9 @@ public:
         return fut.get();  // rethrows if the forward failed
     }
 
-    long Launches() const { return launches_.load(); }
+    long long Launches() const { return launches_.load(); }
     double MeanBatchRows() const {
-        long l = launches_.load();
+        long long l = launches_.load();
         return l ? static_cast<double>(rows_.load()) / l : 0.0;
     }
 
@@ -170,8 +170,9 @@ private:
     std::condition_variable cv_;
     bool stop_ = false;
     std::thread worker_;
-    std::atomic<long> launches_{0};
-    std::atomic<long> rows_{0};
+    // long long: Windows long is 32-bit and total rows overflow it
+    std::atomic<long long> launches_{0};
+    std::atomic<long long> rows_{0};
 };
 
 class ServedBackend : public InferenceBackend {
